- changed strides type from size_t to long in order to allow negative strides
  this would allow decreasing ordering and in-place axes flips.
  However a complete support for negative strides would probably need to
handle an offset in allocation (voxel 0 is not the beginning of the data
block)

- set a Referential in Volume
- beginning of implementation fox axes flipping (WIP)

+++ aimsalgo/src/aimsalgo/distancemap/chamfermask.cc
@@ -470,7 +470,7 @@ static void AimsFillMaskOffset(AimsDistmapMask& mask,
   ASSERT( vol->getSizeT() == 1 );
 
   AimsDistmapMaskPoint *ptr;
-  vector<size_t> strides = vol->getStrides();
+  vector<long> strides = vol->getStrides();
 
   ptr = mask.FirstPoint;
   for (int l=mask.Length;l>0;l--)
