+++ pyaims/src/libpyaims/vector/numpyarrayfunc.h
@@ -42,9 +42,9 @@ namespace aims
   PyObject* initNumpyArray( PyObject*, PyArray_Descr *numType, int ndim,
                             int* dims,
                             char* buffer, bool xyzorder = true,
-                            size_t *strides = 0 );
+                            long *strides = 0 );
   void resizeNumpyArray( PyObject*, int ndim, int* dims, char* buffer,
-                         size_t *strides = 0 );
+                         long *strides = 0 );
 }
 
 #endif
